Implantação do verbo httpPatch utilizando DTOs para não expor o modelo da entidade da camada de domínio.

+++ APICatalogo/Models/Produto.cs
@@ -21,7 +21,7 @@ public class Produto
     [Required]
     [StringLength(300)]
     public string ImagemUrl { get; set; }
-    public float Estoque { get; set; }
+    public int Estoque { get; set; }
     public DateTime DataCadastro { get; set; }
 
     public int CategoriaId { get; set; }
